Converted Observatory class-tree element
R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/2212253002 .

+++ runtime/observatory/lib/src/elements/containers/virtual_collection.dart
@@ -24,7 +24,7 @@ class VirtualCollectionElement extends HtmlElement implements Renderable {
   VirtualCollectionUpdateCallback _update;
   double _itemHeight;
   int _top;
-  int _height;
+  double _height;
   List _items;
   StreamSubscription _onScrollSubscription;
   StreamSubscription _onResizeSubscription;
